Refactor Streamlit app structure; modularize file upload, data preview, model training, and download sections into separate files for improved maintainability and readability
